6. Zigzag Conversion
